backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0
python-multipart==0.0.6

# OCR e Processamento de Imagens
pytesseract==0.3.10
easyocr==1.7.0
pdf2image==1.16.3
pdfplumber==0.10.3
Pillow==10.1.0
opencv-python==4.8.1.78

# LLM e IA
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2

# Utilitários
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
jinja2==3.1.2

# Banco de Dados
sqlalchemy==2.0.23

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1

# PostgreSQL
psycopg2-binary==2.9.9
alembic==1.12.1
